+++ src/app/component-library/catalog/catalog.module.ts
@@ -12,6 +12,7 @@ import {JsonDialogModule} from '../json-dialog/json-dialog.module';
 import {PipesAndDirectivesModule} from '../pipes-and-directives/pipes-and-directives.module';
 import {PropertyGridModule} from '../property-grid/property-grid.module';
 import {UiElementsModule} from '../ui-elements/ui-elements.module';
+import {AssetCardTagListComponent} from './asset-card-tag-list/asset-card-tag-list.component';
 import {AssetDetailDialogDataService} from './asset-detail-dialog/asset-detail-dialog-data.service';
 import {AssetDetailDialogComponent} from './asset-detail-dialog/asset-detail-dialog.component';
 import {AssetDetailDialogService} from './asset-detail-dialog/asset-detail-dialog.service';
@@ -22,8 +23,10 @@ import {ContractOfferIconComponent} from './contract-offer-icon/contract-offer-i
 import {ContractOfferMiniListComponent} from './contract-offer-mini-list/contract-offer-mini-list.component';
 import {DataOfferCardsComponent} from './data-offer-cards/data-offer-cards.component';
 import {IconWithOnlineStatusComponent} from './icon-with-online-status/icon-with-online-status.component';
+import {SmallIconWithOnlineStatusText} from './small-icon-with-online-status-text/small-icon-with-online-status-text.component';
 import {TransferHistoryMiniListComponent} from './transfer-history-mini-list/transfer-history-mini-list.component';
 import {TruncatedShortDescription} from './truncated-short-description/truncated-short-description.component';
+import {ViewSelectionComponent} from './view-selection/view-selection.component';
 
 @NgModule({
   imports: [
@@ -39,6 +42,8 @@ import {TruncatedShortDescription} from './truncated-short-description/truncated
     MatProgressSpinnerModule,
     MatDialogModule,
     MatTooltipModule,
+    MatIconModule,
+    MatChipsModule,
 
     // Features
     JsonDialogModule,
@@ -55,6 +60,9 @@ import {TruncatedShortDescription} from './truncated-short-description/truncated
     IconWithOnlineStatusComponent,
     MarkdownDescriptionComponent,
     TruncatedShortDescription,
+    ViewSelectionComponent,
+    SmallIconWithOnlineStatusText,
+    AssetCardTagListComponent,
   ],
   exports: [
     AssetDetailDialogComponent,
@@ -64,6 +72,9 @@ import {TruncatedShortDescription} from './truncated-short-description/truncated
     TransferHistoryMiniListComponent,
     IconWithOnlineStatusComponent,
     TruncatedShortDescription,
+    ViewSelectionComponent,
+    SmallIconWithOnlineStatusText,
+    AssetCardTagListComponent,
   ],
   providers: [
     AssetPropertyGridGroupBuilder,