+++ src/app/component-library/catalog/view-selection/view-mode-enum.ts
@@ -1,0 +1,4 @@
+export enum ViewModeEnum {
+  GRID,
+  LIST,
+}