+++ src/app/core/utils/local-storage-utils.ts
@@ -1,0 +1,19 @@
+export class LocalStorageUtils {
+  saveData<T>(key: string, value: T) {
+    localStorage.setItem(key, JSON.stringify(value));
+  }
+
+  getData<T>(key: string): T | null {
+    const storedItem = localStorage.getItem(key);
+
+    return storedItem == null ? null : JSON.parse(storedItem);
+  }
+
+  removeData(key: string) {
+    localStorage.removeItem(key);
+  }
+
+  clearData() {
+    localStorage.clear();
+  }
+}