+++ src/app/core/utils/string-utils.ts
@@ -15,3 +15,7 @@ export function everythingAfter(separator: string, s: string): string {
   const index = s.indexOf(separator);
   return index === -1 ? '' : s.substring(index + separator.length);
 }
+
+export function capitalize(s: string) {
+  return s.charAt(0).toUpperCase() + s.slice(1);
+}