+++ src/app/routes/broker-ui/catalog-page/catalog-page/catalog-page.component.ts
@@ -5,8 +5,10 @@ import {BehaviorSubject, Subject} from 'rxjs';
 import {filter, map, takeUntil} from 'rxjs/operators';
 import {Store} from '@ngxs/store';
 import {CatalogPageSortingItem} from '@sovity.de/broker-server-client';
+import {LocalStoredValue} from 'src/app/core/utils/local-stored-value';
 import {AssetDetailDialogDataService} from '../../../../component-library/catalog/asset-detail-dialog/asset-detail-dialog-data.service';
 import {AssetDetailDialogService} from '../../../../component-library/catalog/asset-detail-dialog/asset-detail-dialog.service';
+import {ViewModeEnum} from '../../../../component-library/catalog/view-selection/view-mode-enum';
 import {BrokerServerApiService} from '../../../../core/services/api/broker-server-api.service';
 import {FilterBoxItem} from '../filter-box/filter-box-item';
 import {FilterBoxVisibleState} from '../filter-box/filter-box-visible-state';
@@ -29,6 +31,11 @@ export class CatalogPageComponent implements OnInit, OnDestroy {
   state!: CatalogPageStateModel;
   searchText = new FormControl('');
   sortBy = new FormControl<CatalogPageSortingItem | null>(null);
+  viewModeEnum = ViewModeEnum;
+  viewMode = new LocalStoredValue<ViewModeEnum>(
+    ViewModeEnum.GRID,
+    'brokerui.viewMode',
+  );
   private fetch$ = new BehaviorSubject(null);
 
   // only tracked to prevent the component from resetting